+++ mvc/src/main/java/com/interface21/webmvc/servlet/mvc/DispatcherServlet.java
@@ -1,4 +1,4 @@
-package com.techcourse;
+package com.interface21.webmvc.servlet.mvc;
 
 import com.interface21.webmvc.servlet.ModelAndView;
 import com.interface21.webmvc.servlet.View;
@@ -24,7 +24,7 @@ public class DispatcherServlet extends HttpServlet {
 
     @Override
     public void init() {
-        requestHandlerMappings = new RequestHandlerMappings(getClass().getPackageName());
+        requestHandlerMappings = new RequestHandlerMappings();
         handlerAdapters = new HandlerAdapters();
     }
